+++ src/FunFair.Test.Common.Tests/DependencyInjectionTests.cs
@@ -67,6 +67,9 @@ public sealed class DependencyInjectionTests : DependencyInjectionTestsBase
     private static bool IsProxyObject(string fullTypeName)
     {
         return StringComparer.Ordinal.Equals(x: fullTypeName, y: "Castle.Proxies.ObjectProxy")
-            || fullTypeName.StartsWith(value: "Castle.Proxies.ObjectProxy_", comparisonType: StringComparison.Ordinal);
+            || fullTypeName.StartsWith(
+                value: "Castle.Proxies.ObjectProxy_",
+                comparisonType: StringComparison.Ordinal
+            );
     }
 }