+++ src/FunFair.Test.Common.Tests/ExampleBenchmarks.cs
@@ -1,11 +1,15 @@
-﻿using System.Collections.Generic;
+using System.Collections.Generic;
 using System.Diagnostics.CodeAnalysis;
 using System.Linq;
 using BenchmarkDotNet.Attributes;
 
 namespace FunFair.Test.Common.Tests;
 
-[SuppressMessage(category: "FunFair.CodeAnalysis", checkId: "FFS0012:Make Sealed", Justification = "Benchmarks")]
+[SuppressMessage(
+    category: "FunFair.CodeAnalysis",
+    checkId: "FFS0012:Make Sealed",
+    Justification = "Benchmarks"
+)]
 public class ExampleBenchmarks
 {
     private readonly IReadOnlyList<int> _items = [.. Enumerable.Range(start: 0, count: 100)];