+++ src/FunFair.Test.Common/Extensions/TestLoggerExtension.cs
@@ -12,7 +12,12 @@ public static class TestLoggerExtension
         "CA1873: Evaluation of this argument may be expensive and unnecessary if logging is disabled",
         Justification = "This is a unit test assembly - not so worried about performance"
     )]
-    public static void Received(this ILogger logger, LogLevel logLevel, string message, int received = 1)
+    public static void Received(
+        this ILogger logger,
+        LogLevel logLevel,
+        string message,
+        int received = 1
+    )
     {
         logger
             .Received(received)